Run WinRT '[RelayCommand]' tests for all Roslyn-s
