+++ tests/CommunityToolkit.Mvvm.SourceGenerators.Roslyn4110.UnitTests/Test_SourceGeneratorsDiagnostics.cs
@@ -504,89 +504,6 @@ partial class Test_SourceGeneratorsDiagnostics
     }
 
     [TestMethod]
-    public async Task WinRTRelayCommandIsNotGeneratedBindableCustomPropertyCompatibleAnalyzer_NotTargetingWindows_DoesNotWarn()
-    {
-        const string source = """
-            using CommunityToolkit.Mvvm.ComponentModel;
-            using CommunityToolkit.Mvvm.Input;
-            
-            namespace MyApp
-            {
-                public partial class SampleViewModel : ObservableObject
-                {            
-                    [RelayCommand]
-                    private void DoStuff()
-                    {
-                    }
-                }
-            }
-            """;
-
-        await CSharpAnalyzerWithLanguageVersionTest<WinRTRelayCommandIsNotGeneratedBindableCustomPropertyCompatibleAnalyzer>.VerifyAnalyzerAsync(
-            source,
-            LanguageVersion.CSharp12,
-            editorconfig: []);
-    }
-
-    [TestMethod]
-    public async Task WinRTRelayCommandIsNotGeneratedBindableCustomPropertyCompatibleAnalyzer_TargetingWindows_DoesNotWarn()
-    {
-        const string source = """
-            using CommunityToolkit.Mvvm.ComponentModel;
-            using CommunityToolkit.Mvvm.Input;
-            
-            namespace MyApp
-            {
-                public partial class SampleViewModel : ObservableObject
-                {            
-                    [RelayCommand]
-                    private void DoStuff()
-                    {
-                    }
-                }
-            }
-            """;
-
-        await CSharpAnalyzerWithLanguageVersionTest<WinRTRelayCommandIsNotGeneratedBindableCustomPropertyCompatibleAnalyzer>.VerifyAnalyzerAsync(
-            source,
-            LanguageVersion.CSharp12,
-            editorconfig: [("_MvvmToolkitIsUsingWindowsRuntimePack", true)]);
-    }
-
-    [TestMethod]
-    public async Task WinRTRelayCommandIsNotGeneratedBindableCustomPropertyCompatibleAnalyzer_TargetingWindows_Bindable_Warns()
-    {
-        const string source = """
-            using System;
-            using CommunityToolkit.Mvvm.ComponentModel;
-            using CommunityToolkit.Mvvm.Input;
-            using WinRT;
-            
-            namespace MyApp
-            {
-                [GeneratedBindableCustomProperty]
-                public partial class SampleViewModel : ObservableObject
-                {            
-                    [{|MVVMTK0046:RelayCommand|}]
-                    private void DoStuff()
-                    {
-                    }
-                }
-            }
-
-            namespace WinRT
-            {
-                public class GeneratedBindableCustomPropertyAttribute : Attribute;
-            }
-            """;
-
-        await CSharpAnalyzerWithLanguageVersionTest<WinRTRelayCommandIsNotGeneratedBindableCustomPropertyCompatibleAnalyzer>.VerifyAnalyzerAsync(
-            source,
-            LanguageVersion.CSharp12,
-            editorconfig: [("_MvvmToolkitIsUsingWindowsRuntimePack", true)]);
-    }
-
-    [TestMethod]
     public async Task WinRTGeneratedBindableCustomPropertyWithBasesMemberAnalyzer_NotTargetingWindows_DoesNotWarn()
     {
         const string source = """
